pyproject.toml
[project]
name = "urban-mobility-data-explorer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.3.3",
    "polars>=1.34.0",
    "python-dotenv>=1.1.1",
    "sqlmodel>=0.0.27",
]
